refactor: rename deployservice to deploy
